requirements.txt
absl-py==1.4.0
asttokens==2.2.1
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
cloudpickle==2.2.1
cmake==3.26.4
comm==0.1.2
contourpy==1.0.7
cvxpy==1.3.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
docker-pycreds==0.4.0
ecos==2.0.12
executing==1.2.0
Farama-Notifications==0.0.4
filelock==3.12.0
fonttools==4.38.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
gymnasium==0.28.1
imageio==2.27.0
ipykernel==6.21.2
ipython==8.10.0
jax==0.4.1
jax-jumpy==1.0.0
jaxlib==0.4.1+cuda11.cudnn86
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kaggle==1.5.13
kiwisolver==1.4.4
lit==16.0.5.post0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
minigrid==2.2.1
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opt-einsum==3.3.0
osqp==0.6.3
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.0.0
pqdict==1.2.0
promise==2.3
prompt-toolkit==3.0.37
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.3.0
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
qdldl==0.1.7
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
scs==3.2.3
seaborn==0.12.2
sentry-sdk==1.24.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6
text-unidecode==1.3
tokenizers==0.13.2
torch==1.13.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transforms3d==0.4.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.16
wandb==0.13.5
wcwidth==0.2.6
Werkzeug==2.3.6
